Added published footprint of formation Costmap is updated with smallest circle around formation Switched to Vector2f

+++ fpp_ros/include/fpp_ros/geometry_info/edge_info.h
@@ -4,28 +4,27 @@
 #include <vector>
 #include <list>
 #include <Eigen/Dense>
-#include <shared_ptr.h>
 
 namespace geometry_info
 {
     class EdgeInfo
     {
         public:
-            EdgeInfo(Eigen::Vector2d start_point, Eigen::Vector2d end_point);
+            EdgeInfo(Eigen::Vector2f start_point, Eigen::Vector2f end_point);
 
-            void setStartPoint(Eigen::Vector2d start_point);
-            Eigen::Vector2d getStartPoint();
-            void setEndPoint(Eigen::Vector2d end_point);
-            Eigen::Vector2d getEndPoint();
-            Eigen::Vector2d getEdgeVector();
+            void setStartPoint(Eigen::Vector2f start_point);
+            Eigen::Vector2f getStartPoint();
+            void setEndPoint(Eigen::Vector2f end_point);
+            Eigen::Vector2f getEndPoint();
+            Eigen::Vector2f getEdgeVector();
 
 
         private:
             void calculateEdgeVector();
 
-            std::shared_ptr<Eigen::Vector2d> start_point_;
-            std::shared_ptr<Eigen::Vector2d> end_point_;
+            std::shared_ptr<Eigen::Vector2f> start_point_;
+            std::shared_ptr<Eigen::Vector2f> end_point_;
 
-            std::shared_ptr<Eigen::Vector2d> edge_vector_;
+            std::shared_ptr<Eigen::Vector2f> edge_vector_;
     };
 }
